requirements.txt
asgiref==3.7.2
attrs==23.2.0
boto3==1.34.18
botocore==1.34.18
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
defusedxml==0.7.1
Django==5.0.1
django-allauth==0.61.0
django-debug-toolbar==4.2.0
django-rest-knox==4.2.0
django-webpack-loader==3.0.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
dynaconf==3.2.4
hvac==2.1.0
idna==3.6
inflection==0.5.1
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
knox==0.1.14
loguru==0.7.2
MarkupSafe==2.1.3
oauthlib==3.2.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
s3transfer==0.10.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.7
validators==0.22.0
